requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
python-dotenv
faiss-cpu
pandas
PyPDF2
python-multipart
pydantic
httpx
beautifulsoup4
supabase
lxml
Pillow 